Develop #2

+++ src/controllers/customer-address.controller.ts
@@ -1,14 +1,14 @@
 import CustomerAddressRepository from "../repositories/customer-address.repository";
 
 
-export default class CustomerAddressConntroller{
+export default class CustomerAddressController{
     constructor() {
     }
     repository = new CustomerAddressRepository();
 
     get = async(req: any, res: any, next: any) => {
         try {
-            var data = await this.repository.getAll()
+            const data = await this.repository.getAll();
             res.status(200).send(data)
           } catch (erro) {
             res.status(500).send({
@@ -20,7 +20,13 @@ export default class CustomerAddressConntroller{
     getById = async(req: any, res: any, next: any) => {
         let id = req.params.id
         try {
-            var data = await this.repository.getById(id)
+            const data = await this.repository.getById(id);
+            if(data === null){
+              res.status(404).send({
+                 error_code: 'CUSTOMER ADDRESS NOT FOUND',
+                 error_description: 'Customer Address not found'
+              })
+            }
             res.status(200).send(data)
           } catch (erro) {
             res.status(500).send({
@@ -32,7 +38,7 @@ export default class CustomerAddressConntroller{
     post = async(req: any, res: any, next: any) => {
         let id = req.params.id
         try {
-            var data = await this.repository.create(req.body)
+            const data = await this.repository.create(req.body)
             res.status(200).send(data)
           } catch (erro) {
             res.status(500).send({
@@ -44,7 +50,14 @@ export default class CustomerAddressConntroller{
     put = async(req: any, res: any, next: any) => {
         let id = req.params.id
         try {
-            var data = await this.repository.update(id, req.body)
+            const data = await this.repository.update(id, req.body);
+            if(data[0] === 0){
+              res.status(404).send({
+                 error_code: 'CUSTOMER ADDRESS NOT FOUND',
+                 error_description: 'Customer Address not found'
+              })
+              return
+            }
             res.status(200).send(data)
           } catch (erro) {
             res.status(500).send({
@@ -56,13 +69,21 @@ export default class CustomerAddressConntroller{
     delete = async(req: any, res: any, next: any) => {
         let id = req.params.id
         try {
+            const data = await this.repository.getById(id);
+            if(data == null){
+              res.status(404).send({
+                error_code: 'CUSTOMER ADDRESS NOT FOUND',
+                error_description: 'Customer Address not found'
+             })
+             return
+            }
             await this.repository.deleteById(id)
             res.status(200).send({
-               message: 'Deleted'
+               message: 'Deleted'+data
             })
           } catch (erro) {
             res.status(500).send({
-              message: 'Falha ao processar sua requisição'
+              message: 'Falha ao processar sua requisição'+erro
             })
           }
     }
