+++ src/controllers/measure.controller.ts
@@ -9,9 +9,7 @@ export default class MeasureConntroller {
 
   get = async (req: any, res: any, next: any) => {
     try {
-      console.log('Teste')
-      var data = await this.repository.getAll();
-      console.log("DATA", data,)
+      const data = await this.repository.getAll();
       res.status(200).send(data)
     } catch (erro) {
       res.status(500).send({
@@ -23,8 +21,13 @@ export default class MeasureConntroller {
   getById = async (req: any, res: any, next: any) => {
     let id = req.params.id
     try {
-      var data = await this.repository.getById(id)
-      console.log('DATA')
+      const data = await this.repository.getById(id);
+      if (data === null) {
+        res.status(404).send({
+          error_code: 'MEASURE NOT FOUND',
+          error_description: 'Measure not found'
+        })
+      }
       res.status(200).send(data)
     } catch (erro) {
       res.status(500).send({
@@ -53,16 +56,15 @@ export default class MeasureConntroller {
       }
       if (data.length == 0) {
         res.status(404).send({
-          error_code: 'MEASURE_NOT_FOUND',
+          error_code: 'MEASURES_NOT_FOUND',
           error_description: 'Nenhuma leitura encontrada'
         })
         return;
       }
-      var result = {
+      const result = {
         customer_code: id,
         data
       }
-      console.log('DATA')
       res.status(200).send(result)
     } catch (erro) {
       res.status(500).send({
@@ -72,14 +74,27 @@ export default class MeasureConntroller {
   }
 
   post = async (req: any, res: any, next: any) => {
-    var data: any = []
+    let data: any = []
+    let error_description = '';
     try {
       const { image, customer_code, measure_datetime, measure_type } = req.body;
       let valid = validateReadingMeter({ image, customer_code, measure_datetime, measure_type });
       if (valid.error) {
+        if (valid.error.details[0].message === '\"customer_code\" is not allowed to be empty') {
+          error_description = 'Código de Cliente invalido';
+        }
+        if (valid.error.details[0].message === 'Error code \"measure_type\" is not defined, your custom type is missing the correct messages definition') {
+          error_description = 'Tipo de leitura invalida';
+        }
+        if (valid.error.details[0].message === '\"measure_datetime\" is not allowed to be empty') {
+          error_description = 'Data invalida';
+        }
+        if (valid.error.details[0].message === '\"image\" must be a valid base64 string') {
+          error_description = 'Imagem invalida';
+        }
         res.status(400).send({
           error_code: 'INVALID_DATA',
-          error_description: valid.error.details[0].message
+          error_description: error_description == '' ? valid.error.details[0].message : error_description
         })
         return
       }
@@ -91,10 +106,11 @@ export default class MeasureConntroller {
         for (let i = 0; i < data.length; i++) {
           const month = data[i].measure_datetime.getMonth() + 1;
           const year = data[i].measure_datetime.getFullYear();
-          if (current_year == year && current_month == month) {
+          const type = data[i].measure_type;
+          if (measure_type === type && (current_year == year && current_month == month)) {
             return res.status(409).send({
               error_code: 'DOUBLE_REPORT',
-              error_description: 'Leiture do mês já realizada'
+              error_description: 'Leitura do mês já realizada'
             })
           }
         }
@@ -103,7 +119,7 @@ export default class MeasureConntroller {
 
       const measure_value = parseInt(await analyzeImage(image))
       console.log('Value', measure_value)
-      const uploadResult = uploadImage(image);
+      const uploadResult = uploadImage(customer_code, image);
       if (!uploadResult) {
         return res.status(400).send({
           error_code: 'INVALID_DATA',
@@ -115,9 +131,10 @@ export default class MeasureConntroller {
           customer_code,
           measure_datetime,
           measure_type,
-          measure_value
+          measure_value,
+          has_confirmed: 0
         }
-        var measure_inserted = await this.repository.create(measure)
+        const measure_inserted = await this.repository.create(measure)
         res.status(200).send({
           image_url: uploadResult,
           measure_value,
@@ -136,7 +153,14 @@ export default class MeasureConntroller {
   put = async (req: any, res: any, next: any) => {
     let id = req.params.id
     try {
-      var data = await this.repository.update(id, req.body)
+      const data = await this.repository.update(id, req.body);
+      if (data[0] === 0) {
+        res.status(404).send({
+          error_code: 'MEASURE NOT FOUND',
+          error_description: 'Measure not found'
+        })
+        return
+      }
       res.status(200).send(data)
     } catch (erro) {
       res.status(500).send({
@@ -152,7 +176,7 @@ export default class MeasureConntroller {
       if (valid.error) {
         res.status(400).send({
           error_code: 'INVALID_DATA',
-          error_description: valid.error.details[0].message
+          error_description: 'Dados invalidos'
         })
         return
       }
@@ -160,14 +184,14 @@ export default class MeasureConntroller {
       if (!measure) {
         res.status(404).send({
           error_code: 'MEASURE_NOT_FOUND',
-          error_description: 'Leitura do mês não foi encontrada'
+          error_description: 'Leitura não encontrada'
         })
         return
       }
       if (measure.has_confirmed == 1) {
         res.status(409).send({
           error_code: 'CONFIRMATION_DUPLICATE',
-          error_description: 'Leitura do mês já foi realizada'
+          error_description: 'Leitura do mês já realizada'
         })
         return
       }
@@ -184,13 +208,21 @@ export default class MeasureConntroller {
   delete = async (req: any, res: any, next: any) => {
     let id = req.params.id
     try {
+      const data = await this.repository.getById(id);
+      if (data == null) {
+        res.status(404).send({
+          error_code: 'MEASURE NOT FOUND',
+          error_description: 'Measure not found'
+        })
+        return
+      }
       await this.repository.deleteById(id)
       res.status(200).send({
         message: 'Deleted'
       })
     } catch (erro) {
       res.status(500).send({
-        message: 'Falha ao processar sua requisição'
+        message: 'Falha ao processar sua requisição' + erro
       })
     }
   }
